+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Added CloudStorageTestCaseMixin to take care of wiping data between tests
 - Upgraded the codebase to Python 3
-- Added support for Django 2.2, 3.0 and 3.1
+- Added support for Django up until 4.1
 - Switched to running the tests using tox (remove testapp etc.)
 - Replaced memcache with FileBasedCache (or Memorystore for Redis)
 - Replace storage implementation using the new python3 client
@@ -46,6 +46,7 @@
 ### Bug fixes:
 
 - Made pagination cache keys deterministic
+- Fix issue where `defer_iteration_with_finalize` would always use the default connection and ignore queryset `using`.
 
 ## v1.0.1 (bug fix release)